+++ src/DependencyInjection/CompilerPass/RoutingResolverCompilerPass.php
@@ -15,8 +15,14 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 
+/**
+ * `config/routing.yaml` defines a `routing.resolver` service (which is a Config LoaderResolver).
+ * `config/services/routing/routing_loader.yaml` defines services with the `routing.loader` tag.
+ * This adds them to the routing's config loader resolver.
+ */
 class RoutingResolverCompilerPass implements CompilerPassInterface
 {
+    #[\Override]
     public function process(ContainerBuilder $container): void
     {
         if (false === $container->hasDefinition('routing.resolver')) {